requirements.txt
# Core dependencies
colorama==0.4.6
groq>=0.11.0
ollama>=0.3.0
openai>=1.50.0
google-generativeai>=0.8.0
pandas>=2.0.0
protobuf>=4.25.0
Requests>=2.32.0
tiktoken>=0.7.0
python-dotenv>=1.0.0

# UI Framework
streamlit>=1.28.0
plotly>=5.18.0

# Additional utilities
openpyxl>=3.1.0

# Production monitoring
psutil>=5.9.0

# Testing (optional, for development)
pytest>=7.0.0
pytest-cov>=4.0.0
